requirements.txt
oauth2client==4.1.3
gspread-asyncio==1.4.0
discord.py==1.7.1
sentry-sdk==1.0.0
tortoise-orm==0.17.2
asyncpg==0.22.0
python-dotenv==0.17.0
aerich==0.5.3
